+++ common/sgx/verifier.c
@@ -16,6 +16,7 @@
 #include "endorsements.h"
 #include "quote.h"
 #include "report.h"
+#include "tcbinfo.h"
 
 #if !defined(OE_BUILD_ENCLAVE)
 #include "../../host/sgx/sgxquoteprovider.h"
@@ -218,6 +219,7 @@ static oe_result_t _fill_with_known_claims(
     const oe_uuid_t* format_id,
     const uint8_t* report_body,
     const oe_sgx_endorsements_t* sgx_endorsements,
+    const oe_tcb_info_tcb_level_t* platform_tcb_level,
     oe_datetime_t* valid_from,
     oe_datetime_t* valid_until,
     oe_claim_t* claims,
@@ -230,6 +232,8 @@ static oe_result_t _fill_with_known_claims(
     const sgx_report_body_t* sgx_report_body = NULL;
     size_t claims_index = 0;
     bool flag;
+    oe_sgx_tcb_status_t tcb_status =
+        oe_tcb_level_status_to_sgx_tcb_status(&platform_tcb_level->status);
 
     if (claims_length < OE_REQUIRED_CLAIMS_COUNT + OE_SGX_REQUIRED_CLAIMS_COUNT)
         OE_RAISE(OE_INVALID_PARAMETER);
@@ -390,6 +394,22 @@ static oe_result_t _fill_with_known_claims(
 
     if (format_type != SGX_FORMAT_TYPE_LOCAL)
     {
+        // TCB status.
+        OE_CHECK(oe_sgx_add_claim(
+            &claims[claims_index++],
+            OE_CLAIM_TCB_STATUS,
+            sizeof(OE_CLAIM_TCB_STATUS),
+            &tcb_status,
+            sizeof(tcb_status)));
+
+        // TCB date.
+        OE_CHECK(oe_sgx_add_claim(
+            &claims[claims_index++],
+            OE_CLAIM_TCB_DATE,
+            sizeof(OE_CLAIM_TCB_DATE),
+            &platform_tcb_level->tcb_date,
+            sizeof(platform_tcb_level->tcb_date)));
+
         // Validity from.
         OE_CHECK(oe_sgx_add_claim(
             &claims[claims_index++],
@@ -521,6 +541,7 @@ oe_result_t oe_sgx_extract_claims(
     const uint8_t* custom_claims_buffer,
     size_t custom_claims_buffer_size,
     const oe_sgx_endorsements_t* sgx_endorsements,
+    const oe_tcb_info_tcb_level_t* platform_tcb_level,
     oe_datetime_t* valid_from,
     oe_datetime_t* valid_until,
     oe_claim_t** claims_out,
@@ -589,6 +610,7 @@ oe_result_t oe_sgx_extract_claims(
             report_body,
             report_body_size,
             sgx_endorsements,
+            NULL,
             &local_valid_from,
             &local_valid_until));
         valid_from = &local_valid_from;
@@ -601,6 +623,7 @@ oe_result_t oe_sgx_extract_claims(
         format_id,
         report_body,
         sgx_endorsements,
+        platform_tcb_level,
         valid_from,
         valid_until,
         claims,
@@ -658,12 +681,14 @@ oe_result_t oe_sgx_verify_evidence(
     size_t* claims_length)
 {
     oe_result_t result = OE_UNEXPECTED;
+    oe_result_t result_verify_quote = OE_UNEXPECTED;
     oe_datetime_t* time = NULL;
     oe_datetime_t valid_from = {0};
     oe_datetime_t valid_until = {0};
     uint8_t* local_endorsements_buffer = NULL;
     size_t local_endorsements_buffer_size = 0;
     oe_sgx_endorsements_t sgx_endorsements;
+    oe_tcb_info_tcb_level_t platform_tcb_level = {{0}};
     sgx_evidence_format_type_t format_type = SGX_FORMAT_TYPE_UNKNOWN;
     const uint8_t* report_body = NULL;
     size_t report_body_size = 0;
@@ -800,13 +825,31 @@ oe_result_t oe_sgx_verify_evidence(
             &sgx_endorsements));
 
         // Verify the quote now.
-        OE_CHECK(oe_verify_quote_with_sgx_endorsements(
-            report_body,
-            report_body_size,
-            &sgx_endorsements,
-            time,
-            &valid_from,
-            &valid_until));
+        OE_CHECK_NO_TCB_LEVEL(
+            result_verify_quote,
+            oe_verify_quote_with_sgx_endorsements(
+                report_body,
+                report_body_size,
+                &sgx_endorsements,
+                time,
+                &platform_tcb_level,
+                &valid_from,
+                &valid_until));
+
+        // A successful oe_verify_quote_with_sgx_endorsements() returns either
+        // OE_OK, or OE_TCB_LEVEL_INVALID as OE does not terminate verification
+        // due to TCB status not being up to date. So, the verification result
+        // should be consistent with the retrieved TCB status.
+        if ((result_verify_quote == OE_OK) !=
+            (platform_tcb_level.status.fields.up_to_date == 1))
+        {
+            OE_RAISE_MSG(
+                OE_CONSTRAINT_FAILED,
+                "Inconsistent TCB status: verify quote(%s), tcb status(%s)",
+                oe_result_str(result_verify_quote),
+                oe_sgx_tcb_status_str(oe_tcb_level_status_to_sgx_tcb_status(
+                    &platform_tcb_level.status)));
+        }
     }
 
     // Last step is to return the required and custom claims.
@@ -820,6 +863,7 @@ oe_result_t oe_sgx_verify_evidence(
             custom_claims_buffer,
             custom_claims_buffer_size,
             &sgx_endorsements,
+            &platform_tcb_level,
             &valid_from,
             &valid_until,
             claims,
@@ -840,7 +884,7 @@ oe_result_t oe_sgx_verify_evidence(
         }
     }
 
-    result = OE_OK;
+    result = format_type == SGX_FORMAT_TYPE_LOCAL ? OE_OK : result_verify_quote;
 
 done:
     if (local_endorsements_buffer)