+++ include/openenclave/internal/sgx/plugin.h
@@ -38,6 +38,7 @@ typedef enum _sgx_evidence_format_type_t
  * format_type has the right value.
  * @param[in] custom_claims_buffer_size The size of the custom_claims buffer.
  * @param[in] sgx_endorsements Pointer to the endorsements buffer.
+ * @param[in] platform_tcb_level Pointer to the platform tcb info.
  * @param[in] valid_from Pointer to the datetime from which the evidence and
  * endorsements are valid.
  * @param[in] valid_until Pointer to the datetime at which the evidence and
@@ -50,6 +51,7 @@ typedef enum _sgx_evidence_format_type_t
  * @retval An appropriate error code on failure.
  */
 struct _oe_sgx_endorsements_t;
+struct _oe_tcb_info_tcb_level;
 oe_result_t oe_sgx_extract_claims(
     const sgx_evidence_format_type_t format_type,
     const oe_uuid_t* format_id,
@@ -58,6 +60,7 @@ oe_result_t oe_sgx_extract_claims(
     const uint8_t* custom_claims_buffer,
     size_t custom_claims_buffer_size,
     const struct _oe_sgx_endorsements_t* sgx_endorsements,
+    const struct _oe_tcb_info_tcb_level* platform_tcb_level,
     oe_datetime_t* valid_from,
     oe_datetime_t* valid_until,
     oe_claim_t** claims_out,